fix tests for delete threads

+++ app/Http/Controllers/ThreadsController.php
@@ -124,14 +124,18 @@ class ThreadsController extends Controller
 	 */
 	public function destroy( Channel $channel, Thread $thread )
 	{
+		if ( $thread->user_id != auth()->id() ) {
+			abort(403, 'Ypu do not have permission to do this');
+		}
+
 		$thread->replies()->delete();
 		$thread->delete();
 
-		if(request()->wantsJson()) {
-			return response([], 204);
+		if ( request()->wantsJson() ) {
+			return response( [], 204 );
 		}
 
-		return redirect('/threads');
+		return redirect( '/threads' );
 	}
 
 	/**
